polyintel-main/requirements.txt
fastapi==0.115.7
uvicorn==0.38.0
pydantic==2.12.4
pydantic-core==2.41.5
httpx==0.28.1
starlette==0.45.3
requests==2.32.5
python-dotenv==1.2.1
spoon-ai-sdk 
pinecone
web3
solana
nest_asyncio
anthropic
tenacity

# Optional: Polymarket CLOB (if trading)
#py-clob-client>=0.1.0
elevenlabs==2.24.0